requirements.txt
gunicorn==19.7.1
backcall==0.2.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.6.1
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.19
traitlets==4.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
Flask-WTF==0.14.3
email-validator==1.1.1
phonenumbers==8.12.9
flask-script==2.0.6
